move write_docking_files method to DockingBase class
